Test.Aeson.GenericSpecs

Description

This package provides tools for testing Aeson serialization.

Test that ToJSON and FromJSON instances are isomorphic.

Alert you when unexpected changes in Aeson serialization occur.

Record JSON formatting of Haskell types.

Arbitrary testing

goldenSpecs :: (Typeable a, Arbitrary a, ToJSON a, FromJSON a) => Settings -> Proxy a -> Spec

Tests to ensure that JSON encoding has not unintentionally changed. This could be caused by the following:

A type's instances of ToJSON or FromJSON have changed.

Selectors have been edited, added or deleted.

You have changed version of Aeson the way Aeson serialization has changed works.

If you run this function and the golden files do not exist, it will create them for each constructor. It they do exist, it will compare with golden file if it exists. Golden file encodes json format of a type. It is recommended that you put the golden files under revision control to help monitor changes.

roundtripSpecs :: forall a. (Typeable a, Arbitrary a, ToJSON a, FromJSON a) => Proxy a -> Spec

A roundtrip test to check whether values of the given type can be successfully converted to JSON and back to a Haskell value.

roundtripSpecs will

create random values (using Arbitrary),

convert them into JSON (using ToJSON),

read them back into Haskell (using FromJSON) and

make sure that the result is the same as the value it started with (using Eq).

roundtripAndGoldenSpecs :: forall a. (Arbitrary a, ToJSON a, FromJSON a, Typeable a) => Proxy a -> Spec

run roundtrip and golden test for a type. sampleSize is used only when creating the golden file. When it is compared, the sampleSize is derived from the file.

ToADTArbitrary testing

goldenADTSpecs :: forall a. (ToADTArbitrary a, Eq a, Show a, ToJSON a, FromJSON a) => Settings -> Proxy a -> Spec

Tests to ensure that JSON encoding has not unintentionally changed. This could be caused by the following:

A type's instances of ToJSON or FromJSON have changed.

Selectors have been edited, added or deleted.

You have changed version of Aeson the way Aeson serialization has changed works.

If you run this function and the golden files do not exist, it will create them for each constructor. It they do exist, it will compare with golden file if it exists. Golden file encodes json format of a type. It is recommended that you put the golden files under revision control to help monitor changes.

roundtripADTSpecs :: forall a. (Arbitrary a, ToADTArbitrary a, Eq a, Show a, ToJSON a, FromJSON a) => Proxy a -> Spec

A roundtrip test to check whether values of all of constructors of the given type can be successfully converted to JSON and back to a Haskell value.

roundtripADTSpecs will

create random values for each constructor using ToADTArbitrary,

convert them into JSON using ToJSON,

read them back into Haskell using FromJSON and

make sure that the result is the same as the value it started with using Eq.

roundtripAndGoldenSpecsWithSettings :: forall a. (Arbitrary a, ToJSON a, FromJSON a, Typeable a) => Settings -> Proxy a -> Spec

roundtripAndGoldenSpecs with custom settings.

roundtripAndGoldenADTSpecs :: forall a. (Arbitrary a, ToADTArbitrary a, Eq a, Show a, ToJSON a, FromJSON a) => Proxy a -> Spec

run roundtrip and golden tests for all constructors of a type. sampleSize is used only when creating the golden files. When they are compared, the sampleSize is derived from the file.

roundtripAndGoldenADTSpecsWithSettings :: forall a. (Arbitrary a, ToADTArbitrary a, Eq a, Show a, ToJSON a, FromJSON a) => Settings -> Proxy a -> Spec

roundtripAndGoldenADTSpecs with custom settings.

Make Files

mkGoldenFileForType :: forall a. (ToJSON a, ToADTArbitrary a) => Int -> Proxy a -> FilePath -> IO ()

Make a Golden File for the Proxy of a type if the file does not exist.

Util

shouldBeIdentity :: (Eq a, Show a, Arbitrary a) => Proxy a -> (a -> IO a) -> Property

hspec style combinator to easily write tests that check the a given operation returns the same value it was given, e.g. roundtrip tests.

data GoldenDirectoryOption

A custom directory name or a preselected directory name.

Constructors

CustomDirectoryName String

GoldenDirectory

data Settings

Constructors

Settings

Fields

goldenDirectoryOption :: GoldenDirectoryOption

use a custom directory name or use the generic "golden" directory.

useModuleNameAsSubDirectory :: Bool

If true, use the module name in the file path, otherwise ignore it.

sampleSize :: Int

How many instances of each type you want. If you use ADT versions than it will use the sample size for each constructor.

comparisonFile :: ComparisonFile

Whether to create a separate comparison file or ovewrite the golden file.

randomMismatchOption :: RandomMismatchOption

Whether to output a warning or fail the test when the random seed produces different values than the values in the golden file.

defaultSettings :: Settings

The default settings for general use cases.

re-exports

data Proxy (t :: k)

Proxy is a type that holds no data, but has a phantom parameter of arbitrary type (or even kind). Its use is to provide type information, even though there is no value available of that type (or it may be too costly to create one).

Historically, Proxy :: Proxy a is a safer alternative to the undefined :: a idiom.

>>> Proxy :: Proxy (Void, Int -> Int)
Proxy

Proxy can even hold types of higher kinds,

>>> Proxy :: Proxy Either
Proxy

>>> Proxy :: Proxy Functor
Proxy

>>> Proxy :: Proxy complicatedStructure
Proxy

Constructors

Proxy