Test.Aeson.Internal.ADT.RoundtripSpecs

Description

Internal module, use at your own risk.

Documentation

roundtripADTSpecs :: forall a. (Arbitrary a, ToADTArbitrary a, Eq a, Show a, ToJSON a, FromJSON a) => Proxy a -> Spec

A roundtrip test to check whether values of all of constructors of the given type can be successfully converted to JSON and back to a Haskell value.

roundtripADTSpecs will

create random values for each constructor using ToADTArbitrary,

convert them into JSON using ToJSON,

read them back into Haskell using FromJSON and

make sure that the result is the same as the value it started with using Eq.

genericAesonRoundtripADTWithNote :: forall a. (ToADTArbitrary a, Eq a, Show a, Arbitrary a, ToJSON a, FromJSON a) => Proxy a -> Maybe String -> Spec

Same as roundtripADTSpecs but has the option of passing a note to the describe function.