Test.Aeson.Internal.RandomSamples

Description

Internal module, use at your own risk.

Documentation

data RandomSamples a

RandomSamples, using a seed allows you to replicate an arbitrary. By storing the seed and the samples (previously produced arbitraries), we can try to reproduce the same samples by generating the arbitraries with a seed.

Constructors

RandomSamples

Fields

seed :: Int32

samples :: [a]

setSeed :: Int -> Gen a -> Gen a

Apply the seed.

readSeed :: ByteString -> IO Int32

Reads the seed without looking at the samples.

readSampleSize :: ByteString -> IO Int

Read the sample size.