Test.Aeson.Internal.Utils

Description

Documentation

data ComparisonFile

Option to indicate whether to create a separate comparison file or overwrite the golden file. A separate file allows you to use diff to compare. Overwriting allows you to use source control tools for comparison.

Constructors

FaultyFile
Create a new faulty file when tests fail

OverwriteGoldenFile
Overwrite the golden file when tests fail

data RandomMismatchOption

Option indicating whether to fail tests when the random seed does not produce the same values as in the golden file. Default is to output a warning.

Constructors

RandomMismatchWarning
Only output a warning when the random seed does not produce the same values

RandomMismatchError
Fail the test when the random seed does not produce the same value

data Settings

Constructors

Settings

Fields

goldenDirectoryOption :: GoldenDirectoryOption

use a custom directory name or use the generic "golden" directory.

useModuleNameAsSubDirectory :: Bool

If true, use the module name in the file path, otherwise ignore it.

sampleSize :: Int

How many instances of each type you want. If you use ADT versions than it will use the sample size for each constructor.

comparisonFile :: ComparisonFile

Whether to create a separate comparison file or ovewrite the golden file.

randomMismatchOption :: RandomMismatchOption

Whether to output a warning or fail the test when the random seed produces different values than the values in the golden file.

data GoldenDirectoryOption

A custom directory name or a preselected directory name.

Constructors

CustomDirectoryName String

GoldenDirectory

defaultSettings :: Settings

The default settings for general use cases.

addBrackets :: String -> String

put brackets around a String.

shouldBeIdentity :: (Eq a, Show a, Arbitrary a) => Proxy a -> (a -> IO a) -> Property

hspec style combinator to easily write tests that check the a given operation returns the same value it was given, e.g. roundtrip tests.

checkAesonEncodingEquality :: forall a. (ToJSON a, FromJSON a) => JsonShow a -> Bool

This function will compare one JSON encoding to a subsequent JSON encoding, thus eliminating the need for an Eq instance

aesonDecodeIO :: FromJSON a => ByteString -> IO a

run decode in IO, if it returns Left then throw an error.

data AesonDecodeError

Constructors

AesonDecodeError String

newtype JsonShow a

Used to eliminate the need for an Eq instance

Constructors

JsonShow a

newtype TopDir

Constructors

TopDir

Fields

unTopDir :: FilePath

newtype ModuleName

Constructors

ModuleName

Fields

unModuleName :: FilePath

newtype TypeName

Constructors

TypeName

Fields

unTypeName :: FilePath

data TypeNameInfo a

Constructors

TypeNameInfo

Fields

typeNameTopDir :: TopDir

typeNameModuleName :: Maybe ModuleName

typeNameTypeName :: TypeName

mkTypeNameInfo :: forall a. Arbitrary a => Typeable a => Settings -> Proxy a -> IO (TypeNameInfo a)

encodePrettySortedKeys :: ToJSON a => a -> ByteString